+++ src/main/java/com/pocketserver/net/packets/ping/PongPacket.java
@@ -16,7 +16,7 @@ public class PongPacket extends OutPacket {
 
     @Override
     public DatagramPacket encode(DatagramPacket buf) {
-    	buf.content().writeInt(getPacketID());
+    	buf.content().writeByte(getPacketID());
     	buf.content().writeLong(identifier);
         return buf;
     }
